Update README.rst
Add python3 -m pip install scipy==1.8, because the new version, of scipy (==1.9) changes coo_array. To avoid error, version 1.8 of scipy must be installed. This precision is taken in the github but not yet in the pip version. So for the quick pip installation, it needed to install manually scipy==1.8.